cookbook/rag_examples/requirements.txt
mirascope>=0.2.0,<1.0.0
numpy>=1
pandas>=1.2.3
tiktoken
pinecone-client
python-dotenv
